requirements.txt
abnf==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.10.0
asttokens==2.4.1
attrs==23.2.0
autobahn==25.10.2
Automat==25.4.16
bitarray==2.9.2
black==24.2.0
cachetools==6.2.1
cbor2==5.6.2
certifi==2024.2.2
cffi==2.0.0
cfgv==3.4.0
channels==4.3.1
charset-normalizer==3.3.2
ckzg==2.1.5
click==8.1.7
constantly==23.10.4
cryptography==46.0.3
cytoolz==0.12.3
daphne==4.2.1
dataclassy==0.11.1
distlib==0.4.0
Django==5.2
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
ecdsa==0.19.1
eip712==0.2.4
eth-account==0.10.0
eth-brownie==1.20.6
eth-event==1.2.5
eth-hash==0.6.0
eth-keyfile==0.7.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==3.5.2
eth-utils==2.3.1
eth_abi==5.0.0
execnet==2.0.2
filelock==3.20.0
frozenlist==1.4.1
google-auth==2.42.0
hexbytes==0.3.1
hyperlink==21.0.0
hypothesis==6.27.3
identify==2.6.15
idna==3.6
importlib-metadata==7.0.1
incremental==24.7.2
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
loguru==0.7.3
lru-dict==1.2.0
multidict==6.0.5
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==23.2
parsimonious==0.9.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pre_commit==4.1.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
py==1.11.0
py-solc-ast==1.2.10
py-solc-x==1.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.20.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.17.2
pygments-lexer-solidity==0.7.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==6.2.5
pytest-forked==1.6.0
pytest-xdist==1.34.0
python-dateutil==2.9.0.post0
python-dotenv==0.16.0
pyunormalize==15.1.0
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rlp==4.0.0
rpds-py==0.18.0
rsa==4.9.1
semantic-version==2.10.0
service-identity==24.2.0
setuptools==80.9.0
siwe==4.4.0
siws==0.0.2
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.5.3
toml==0.10.2
toolz==0.12.1
tqdm==4.66.2
Twisted==25.5.0
txaio==25.9.2
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.9.0
urllib3==2.2.1
virtualenv==20.35.3
vvm==0.1.0
vyper==0.3.10
wcwidth==0.2.13
web3==6.15.1
websockets==12.0
wheel==0.42.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
zope.interface==8.0.1
